move zod schema to ott common and fix lint
